feat: switch to message relays
