+++ src/contents/yt.ts
@@ -1,7 +1,6 @@
 import type { PlasmoCSConfig } from 'plasmo'
 
-import { logError, logLine } from '~helpers/logging'
-import type { MarkNotifReadEventDetail, YTData } from '~interfaces'
+import type { YTData } from '~interfaces'
 
 export const config: PlasmoCSConfig = {
   matches: ['*://*.youtube.com/*'],
@@ -52,86 +51,3 @@ const handleNavigateFinish = (event) => {
 }
 
 window.addEventListener('yt-navigate-finish', handleNavigateFinish)
-
-const markNotificationAsRead = async (event: CustomEvent) => {
-  const { xpath, authorizationHeader, ytData, loggingEnabled } =
-    event.detail as MarkNotifReadEventDetail
-  const { authUser, clientVersion, pageId, visitorId } = ytData
-
-  const element: any = document.evaluate(
-    xpath,
-    document,
-    null,
-    XPathResult.FIRST_ORDERED_NODE_TYPE,
-    null,
-  ).singleNodeValue
-  const elementData: any = element?.data
-
-  try {
-    if (
-      !authUser ||
-      !clientVersion ||
-      !visitorId ||
-      !authorizationHeader ||
-      !elementData
-    ) {
-      logLine(
-        'Missing required data',
-        [
-          authUser,
-          clientVersion,
-          pageId,
-          visitorId,
-          authorizationHeader,
-          elementData,
-        ],
-        loggingEnabled,
-      )
-      return
-    }
-
-    const payload = {
-      context: {
-        client: {
-          clientName: 'WEB',
-          clientVersion,
-        },
-      },
-      serializedRecordNotificationInteractionsRequest:
-        elementData.recordClickEndpoint.recordNotificationInteractionsEndpoint
-          .serializedInteractionsRequest,
-    }
-
-    const response = await fetch(
-      'https://www.youtube.com/youtubei/v1/notification/record_interactions?prettyPrint=false',
-      {
-        method: 'POST',
-        headers: {
-          Authorization: `SAPISIDHASH ${authorizationHeader}`,
-          'Content-Type': 'application/json',
-          'X-Origin': 'https://www.youtube.com',
-          'X-Goog-Authuser': authUser,
-          // PageId seems to be only available when you've switched to a different user from the original one.
-          ...(pageId ? { 'X-Goog-PageId': pageId } : {}),
-          'X-Goog-Visitor-Id': visitorId,
-          'X-Youtube-Bootstrap-Logged-In': 'true',
-          'X-Youtube-Client-Name': '1',
-          'X-Youtube-Client-Version': clientVersion,
-        },
-        body: JSON.stringify(payload),
-      },
-    )
-
-    const responseJson = await response.json()
-
-    if (response.ok && responseJson?.success) {
-      logLine('Notification marked as read!')
-    } else {
-      logError('Failed to mark notification as read', [responseJson])
-    }
-  } catch (error) {
-    logError('Failed to mark notification as read', [error])
-  }
-}
-
-window.addEventListener('ytwl-mark-notif-read', markNotificationAsRead)
