+++ src/helpers/logging.ts
@@ -1,27 +1,22 @@
-import { Storage } from '@plasmohq/storage'
+import { sendToBackgroundViaRelay } from '@plasmohq/messaging'
 
-export const loggingEnabled = async (): Promise<boolean> => {
-  const storage = new Storage()
-  const isLogging: boolean = await storage.get('isLogging')
-  return isLogging
+import type { Settings } from '~interfaces'
+
+const loggingEnabled = async (): Promise<boolean> => {
+  const settings: Settings = await sendToBackgroundViaRelay<Settings>({
+    name: 'settings',
+  })
+  return settings.loggingEnabled
 }
 
-export const logLine = async (
-  message: string,
-  optionalParams: any[] = [],
-  _loggingEnabled: boolean | null = null,
-) => {
-  if (_loggingEnabled === true || (await loggingEnabled())) {
+export const logLine = async (message: string, ...optionalParams: any[]) => {
+  if (await loggingEnabled()) {
     console.log(`[YT Watch Later] ${message}`, ...optionalParams)
   }
 }
 
-export const logError = async (
-  message: string,
-  optionalParams: any[] = [],
-  _loggingEnabled: boolean | null = null,
-) => {
-  if (_loggingEnabled === true || (await loggingEnabled())) {
+export const logError = async (message: string, ...optionalParams: any[]) => {
+  if (await loggingEnabled()) {
     console.error(`[YT Watch Later] ${message}`, ...optionalParams)
   }
 }
